Silence which command's standard error output
The GNU which v2.21 command found on CentOS Stream 8 behaves differently
than the which command on Debian based systems such as Ubuntu 20.04. The
GNU which command prints to stderr when the command is not found. This
can cause an error message to be erroneously printed during our test for
a locally installed VS Code binary. Redirecting stderr to /dev/null
silences this error message so our code produces the same behavior for
either version of the which command.